timeout support for <java>
PR: 5299


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@272237 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/ExecTask.java
@@ -88,7 +88,7 @@ public class ExecTask extends Task {
     private File dir;
     protected boolean failOnError = false;
     protected boolean newEnvironment = false;
-    private Integer timeout = null;
+    private Long timeout = null;
     private Environment env = new Environment();
     protected Commandline cmdl = new Commandline();
     private FileOutputStream fos = null;
@@ -104,7 +104,7 @@ public class ExecTask extends Task {
     /**
      * Timeout in milliseconds after which the process will be killed.
      */
-    public void setTimeout(Integer value) {
+    public void setTimeout(Long value) {
         timeout = value;
     }
 
@@ -377,7 +377,7 @@ public class ExecTask extends Task {
         if (timeout == null) {
           return null;
         }
-        return new ExecuteWatchdog(timeout.intValue());
+        return new ExecuteWatchdog(timeout.longValue());
     }
 
     /**
